rename presentation to adminapi
